% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_completed.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates a summary table of the number of fatal accidents by month by year for
a given set of years for a given state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number indicating the state.}

\item{year}{A numeric or interter; should be the full 4 digit year.}
}
\value{
A map of the fatal accidents for a state in a given year.
}
\description{
Creates a summary table of the number of fatal accidents by month by year for
a given set of years for a given state.
}
\details{
An error will result if the years are not given as full 4 digit
  representions or if the indicated packages are not attached.
}
\examples{
fars_map_state(1,2013)
}
