% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_completed.R
\name{fars_read}
\alias{fars_read}
\title{Read a csv file in as a tbl for a particular year of data from the Fatality
Analysis Reporting System.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of a csv file.}
}
\value{
A tbl representation of the data contained in the csv file.
}
\description{
Read a csv file in as a tbl for a particular year of data from the Fatality
Analysis Reporting System.
}
\details{
An error will result if the file name does not exist.
}
\examples{
\dontrun{
fars_read("accident_2013.csv.bz2")
}
}
