% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_completed.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a summary table of the number of fatal accidents by month by year for
a given set of years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric or integer vector of years; should be the full 4 digit
years.}
}
\value{
A table of the number of fatal accidents by year by month for the
  given set of years.
}
\description{
Creates a summary table of the number of fatal accidents by month by year for
a given set of years.
}
\details{
An error will result if the years are not given as full 4 digit
  representions or if the indicated packages are not attached.
}
\examples{
fars_summarize_years(2013:2015)
}
