% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aFunction.R
\name{foo}
\alias{foo}
\title{Foo a random function just as a test}
\usage{
foo(x)
}
\arguments{
\item{x}{An argument of any type}
}
\value{
x That argument right back at you
}
\description{
Foo a random function just as a test
}
