% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions_completed.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name for a paticular year's data file from the Fatality Analysis
Reporting System.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric or interter; should be the full 4 digit year.}
}
\value{
The name of the file from the Fatality Analysis Reporting System for
  that year.
}
\description{
Make a file name for a paticular year's data file from the Fatality Analysis
Reporting System.
}
\details{
An error will result if the year is outside the scope of years
  available as data sets.
}
\examples{
\dontrun{
foo<-fars_read(make_filename(2013))
}
}
